+++ src/app/(public)/resources/_components/add-new-resource-form.tsx
@@ -1,3 +1,0 @@
-export const AddNewResourceForm = () => {
-  return <div>AddNewResourceForm</div>
-}